:children_crossing: when we go to the next step in register focus the first input of this step, fix button stay in loading after result in login page

+++ back/app/controllers/auth_controller.ts
@@ -31,7 +31,10 @@ export default class AuthController {
           .select('image')
           .first()
 
-        if(userImage) {
+        if(
+          userImage &&
+          userImage.image
+        ) {
           const imageLink: string = `./uploads/${userImage.image}`
 
           fs.unlinkSync(imageLink)
